requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.3.8
filelock==3.13.1
keyboard==0.13.5
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
plyer==2.1.0
setuptools==69.1.1
tk==0.1.0
ttkbootstrap==1.10.1
virtualenv==20.25.0
wheel==0.42.0
